feat(app/tabs/Home): Implement Home page for table booking app
Implement the Home page screen for the table booking application. The screen includes two separate text inputs for departure and destination cities, and a search button that does not perform any functionality yet. Additionally, an independent component called LocationInputs has been implemented to handle the city selection process, along with a new component called LocationBottomSheet that is used for the modal display of available cities. For now, the modal is passive, but it can be interacted with for future implementation of city selection functionality. Furthermore, the old Index screen has been replaced with the new Home screen. Lastly, a new type for Location has been added in the src/types/location.ts file.